+++ src/containers/custom/CustomPlaceFilterContainer.test.jsx
@@ -10,11 +10,11 @@ import {
   MemoryRouter,
 } from 'react-router-dom';
 
-import CustomFilterContainer from './CustomFilterContainer';
+import CustomPlaceFilterContainer from './CustomPlaceFilterContainer';
 
 jest.mock('react-redux');
 
-describe('CustomFilterContainer', () => {
+describe('CustomPlaceFilterContainer', () => {
   const mock = jest.fn();
   const dispatch = jest.fn();
 
@@ -25,33 +25,8 @@ describe('CustomFilterContainer', () => {
     useDispatch.mockImplementation(() => dispatch);
 
     useSelector.mockImplementation((selector) => (
-      selector.categoryColor = 'selector',
       selector.placeColor = 'selector',
-      selector.selectedCategory = [ // 양식
-        {
-          "id": "10",
-          "name": "더다이닝랩",
-          "situation": "소개팅",
-          "age": "20대",
-          "place": "홍대/합정",
-          "category": "양식",
-          "priceRange": "3만원 이하",
-          "mood": "none",
-          "2nd-course": "none",
-        },
-        {
-          "id": "36",
-          "name": "보이어",
-          "situation": "데이트",
-          "age": "20대",
-          "place": "성수",
-          "category": "양식",
-          "priceRange": "3만원 이하",
-          "mood": "고급스러운",
-          "2nd-course": "none",
-        },
-      ],
-      selector.selectedPlace = [ // 광화문/종로
+      selector.selectedPlace = [
         {
           "id": "21",
           "name": "갈리나데이지",
@@ -126,7 +101,7 @@ describe('CustomFilterContainer', () => {
 
   const renderCustomFilterContainer = () => render((
     <MemoryRouter>
-      <CustomFilterContainer />
+      <CustomPlaceFilterContainer />
     </MemoryRouter>
   ));
 
@@ -178,37 +153,14 @@ describe('CustomFilterContainer', () => {
       },
     ]
 
-    const uniqCategories = uniqBy(restaurantsData, 'category');
     const uniqPlaces = uniqBy(restaurantsData, 'place');
-
-    it('calls its argument with a non-null argument (1)', () => {
-      const { container } = renderCustomFilterContainer();
-
-      expect(container).toHaveTextContent('무엇을 드시고 싶으세요?');
-
-      uniqCategories.map(category => mock(category))
-      expect(mock).toBeCalled();
-    });
-
-    it('calls its argument with a non-null argument (2)', () => {
+    it('calls its argument with a non-null argument', () => {
       const { container } = renderCustomFilterContainer();
 
       expect(container).toHaveTextContent('어디로 가고 싶나요?');
 
       uniqPlaces.map(place => mock(place))
       expect(mock).toBeCalled();
-    });
-  });
-
-  context('when click "#양식" tag', () => {
-    it('calls dispatch with action : setCategoryFilter', () => {
-      const { getByText } = renderCustomFilterContainer();
-
-      expect(getByText('#양식')).toBeInTheDocument();
-
-      fireEvent.click(getByText('#양식'));
-
-      expect(dispatch).toBeCalled();
     });
   });
